Register beans and tools for monitoring in application.
