+++ src/main/java/edu/kit/datamanager/metastore2/Application.java
@@ -23,14 +23,13 @@ import edu.kit.datamanager.configuration.SearchConfiguration;
 import edu.kit.datamanager.entities.messaging.IAMQPSubmittable;
 import edu.kit.datamanager.metastore2.configuration.ApplicationProperties;
 import edu.kit.datamanager.metastore2.configuration.MetastoreConfiguration;
+import edu.kit.datamanager.metastore2.configuration.MonitoringConfiguration;
 import edu.kit.datamanager.metastore2.configuration.OaiPmhConfiguration;
-import edu.kit.datamanager.metastore2.dao.IDataRecordDao;
-import edu.kit.datamanager.metastore2.dao.IMetadataFormatDao;
-import edu.kit.datamanager.metastore2.dao.ISchemaRecordDao;
-import edu.kit.datamanager.metastore2.dao.IUrl2PathDao;
+import edu.kit.datamanager.metastore2.dao.*;
 import edu.kit.datamanager.metastore2.util.DataResourceRecordUtil;
 import edu.kit.datamanager.metastore2.util.MetadataRecordUtil;
 import edu.kit.datamanager.metastore2.util.MetadataSchemaRecordUtil;
+import edu.kit.datamanager.metastore2.util.MonitoringUtil;
 import edu.kit.datamanager.metastore2.validation.IValidator;
 import edu.kit.datamanager.repo.configuration.DateBasedStorageProperties;
 import edu.kit.datamanager.repo.configuration.IdBasedStorageProperties;
@@ -117,6 +116,8 @@ public class Application {
   @Autowired
   private IMetadataFormatDao metadataFormatDao;
   @Autowired
+  private IIpMonitoringDao ipMonitoringDao;
+  @Autowired
   private List<IValidator> validators;
 
   @Bean
@@ -159,6 +160,11 @@ public class Application {
   @Bean
   public OaiPmhConfiguration oaiPmhConfiguration() {
     return new OaiPmhConfiguration();
+  }
+
+  @Bean
+  public MonitoringConfiguration monitoringConfiguration() {
+    return new MonitoringConfiguration();
   }
 
   @Bean
@@ -288,6 +294,9 @@ public class Application {
     DataResourceRecordUtil.setSchemaConfig(rbc);
     DataResourceRecordUtil.setUrl2PathDao(url2PathDao);
     DataResourceRecordUtil.setAllIdentifiersDao(allIdentifiersDao);
+    MonitoringUtil.setIpMonitoringDao(ipMonitoringDao);
+    MonitoringUtil.setMonitoringConfiguration(monitoringConfiguration());
+
 
     fixBasePath(rbc);
